backend/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "fastapi==0.115.8",
    "httpx==0.28.1",
    "pydantic[email]==2.10.6",
    "pydantic-settings>=2.10.0",
    "pyjwt==2.10.1",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "ruff>=0.12.0",
    "sqlalchemy==2.0.38",
    "uvicorn==0.34.0",
    "python-multipart>=0.0.20",
    "argon2-cffi>=25.1.0",
]
